// pokemon-wird/src/interface/interfaces.ts
export interface PokemonBasic {
  name: string;
  url: string;
}

// export interface PokemonDetailed extends PokemonBasic {
//   id?: number;
//   sprites?: {
//     front_default?: string;
//   };
//   height?: number;
//   weight?: number;
//   types?: Array<{
//     type: {
//       name: string;
//     };
//   }>;
//   // Añade más propiedades según necesites, todas opcionales
// }

export interface PokemonState {
  list: PokemonBasic[];
  detailedList: PokemonDetailed[];
  battleTeam: PokemonDetailed[];
  searchTerm: string;
  loading: boolean;
  selectedPokemon: PokemonDetailed | null;
  error: string | null;
}

export interface PokemonDetailed extends PokemonBasic {
  id: number;
  name: string;
  base_experience: number;
  height: number;
  is_default: boolean;
  order: number;
  weight: number;
  abilities: Ability2[];
  forms: Ability[];
  game_indices: Gameindex[];
  held_items: Helditem[];
  location_area_encounters: string;
  moves: Move[];
  species: Ability;
  sprites: Sprites;
  cries: Cries;
  stats: Stat[];
  types: Type[];
  past_types: Pasttype[];
}
interface Pasttype {
  generation: Ability;
  types: Type[];
}
interface Type {
  slot: number;
  type: Ability;
}
interface Stat {
  base_stat: number;
  effort: number;
  stat: Ability;
}
interface Cries {
  latest: string;
  legacy: string;
}
interface Sprites {
  back_default: string;
  back_female?: any;
  back_shiny: string;
  back_shiny_female?: any;
  front_default: string;
  front_female?: any;
  front_shiny: string;
  front_shiny_female?: any;
  other: Other;
  versions: Versions;
}
interface Versions {
  'generation-i': Generationi;
  'generation-ii': Generationii;
  'generation-iii': Generationiii;
  'generation-iv': Generationiv;
  'generation-v': Generationv;
  'generation-vi': Generationvi;
  'generation-vii': Generationvii;
  'generation-viii': Generationviii;
}
// ...
interface Other {
  dream_world: Dreamworld;
  home: Home;
  'official-artwork': Officialartwork;
  showdown: Showdown;
}
// ...
interface Move {
  move: Ability;
  version_group_details: Versiongroupdetail[];
}
interface Versiongroupdetail {
  level_learned_at: number;
  version_group: Ability;
  move_learn_method: Ability;
}
interface Helditem {
  item: Ability;
  version_details: Versiondetail[];
}
interface Versiondetail {
  rarity: number;
  version: Ability;
}
interface Gameindex {
  game_index: number;
  version: Ability;
}
interface Ability2 {
  is_hidden: boolean;
  slot: number;
  ability: Ability;
}
interface Ability {
  name: string;
  url: string;
}